Finish setting up email registration
